requirements.txt
# 图片压缩工具依赖包

# 图像处理
Pillow>=10.0.0

# 命令行
click>=8.1.0

# 日志（可选）
colorlog>=6.7.0
